Add port_scripts.h for script data resolution on PC port
This new header file defines macros for resolving GBA ROM addresses of script data to native pointers on the PC port. It includes definitions for known script addresses and provides a way to use these addresses in a platform-independent manner. The macros PORT_SCRIPT and ENTITY_SCRIPT facilitate the correct resolution of script data for both GBA and PC builds.

+++ include/color.h
@@ -23,7 +23,7 @@ typedef struct {
     u16 unk_02;
     union SplitWord unk_04;
     union SplitWord unk_08;
-    u32 unk_0C;
+    uintptr_t unk_0C; /* Entity* on PC (was u32 on GBA) */
 } struct_gUnk_020000C0_1;
 
 typedef struct {
